Read event data in parallel to backtest
Remove the `read_data()` calls within the backtest implementations and
replace them with `recv()` calls on a lock-free queue. This avoids the
pause that happened previously when a backtest reaches the end of the
current periods data and begins loading the next file. With this model,
the data for the next period should be available by the time the
previous one finishes.